Added minimized version (along with makefile) for ease of implementation
An example html file has been provided to show just how simple it is to
add roswebkit to an existing webpage. All it requires is a call to
roswebkit(bridgeport, mjpegport) and it will hook into all canvas on the
page that are a class of one of the rosplots.