Added ability to use plots directly as tag names
This allows for a little bit more intuitive html files, at the expense
of doing something that probably isn't supported across all browsers.
Since this already uses websockets and canvases I'm not sure that's an
issue.

Replaces any ROS tags with a canvas and uses the plot as the class name
for example <rosplot> becomes <canvas class=rosplot>