Integrated the build process of roswebkit into ROS itself
Currently only supports rosmake, might expand that later

Honestly the integration is slight at best. The CMakeList simply tells
cmake to use the existing Makefile. While this seems like a stopgap
solution is was not implemented without great effort to make CMake work
with javascript files. CMake has shown to be incredably frustrating to
use with anything that is not c/c++/etc. Simply leaving the build
process in an easily understandable Makefile will let anyone who wants
to use this project in any variant of ROS be able to do so.